[FIX] Display Battery value properly
- Battery value in low battery cases was displayed with two percent 
signs (%%) instead of one
- In Battery History, Battery value was displayed with decimal dot, 
which is kinda useless

+++ app/src/main/java/de/bikebean/app/ui/drawer/history/battery/BatteryHistoryAdapter.java
@@ -54,7 +54,7 @@ class BatteryHistoryAdapter extends HistoryAdapter {
         if (mStates != null) {
             final @NonNull State current = (State) mStates.get(position);
 
-            final @NonNull String batteryStatus = current.getValue() + " %";
+            final @NonNull String batteryStatus = current.getValue().intValue() + " %";
             bHolder.batteryValue.setText(batteryStatus);
             bHolder.batteryValue.setCompoundDrawablesWithIntrinsicBounds(
                     ResourceUtils.getBatteryDrawable(ctx, current.getValue()),
